requirements.txt
djangorestframework==3.15.1
beautifulsoup4==4.12.3
django-cors-headers==4.3.1
bs4==0.0.2
certifi==2024.2.2
Django==4.2.11
decorator
django-jazzmin
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
django-countries
django-redis
boto3==1.21.1
pirate-get==0.4.2
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
IMDbPY==2022.7.9
geoip2==4.8.0
stripe==7.10.0
twilio==9.0.0
pycryptodome==3.20.0
torch==2.2.2 
paypalrestsdk==1.13.3
importlib_resources==6.4.0
moviepy
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
dj-database-url==0.2.1
numpy==1.24.4
pandas==2.0.2
pillow==10.3.0
proglog==0.1.10
psycopg2-binary==2.9.3
psycopg==3.1.8
pydantic==2.7.0
pydantic_core==2.18.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
textstat==0.7.3
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
transformers==4.39.3
